backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy>=2.0
psycopg2-binary==2.9.9
pydantic==2.8.2
httpx==0.27.0
python-dotenv==1.0.1
openai==1.45.0
weaviate-client==4.7.1
python-multipart==0.0.9
alembic==1.13.1
gunicorn==21.2.0
